pyproject.toml
[project]
name = "BlueStacksAgent"
version = "0.2.2"
description = "A Python library for real-time interaction with BlueStacks using scrcpy, minicap, and MediaProjection."
readme = "README.md"
requires-python = ">=3.7, <3.11"
license = { file = "LICENSE" }
dependencies = [
    "scrcpy-client",
    "numpy",
]

[build-system]
requires = [
    "setuptools>=42",
    "wheel"
]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests", "tests.*"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
